chore: move all deploy-tests to one file
