--- packages/@cdktf/cli-core/src/lib/models/__tests__/deploy-machine.test.ts
@@ -1,41 +1,0 @@
-import _ from "lodash";
-import { handleLineReceived } from "../deploy-machine";
-
-describe("handleLineReceived", () => {
-  it("should forward the line", () => {
-    const send = jest.fn();
-    handleLineReceived(send)("foo");
-    expect(send).toHaveBeenCalledWith({
-      type: "LINE_RECEIVED",
-      line: "foo",
-    });
-  });
-
-  it("should hide the line if a variable is missing", () => {
-    const send = jest.fn();
-    const input = `
-        var.my_var
-        Enter a value:
-    `;
-    handleLineReceived(send)(input);
-    expect(send).not.toHaveBeenCalledWith({
-      type: "LINE_RECEIVED",
-      line: input,
-    });
-    expect(send).toHaveBeenCalledWith({
-      type: "VARIABLE_MISSING",
-      variableName: "my_var",
-    });
-  });
-
-  it("should detect destroy all resources in workspace message on windows", () => {
-    const send = jest.fn();
-    const input = `\u001b[25
-         l\u001b[m\n\u001b[1m\u001b[97mDo you really want to destroy all resources in workspace \\"windows-node-test\\"?   \u001b[2
-    `;
-    handleLineReceived(send)(input);
-    expect(send).toHaveBeenCalledWith({
-      type: "REQUEST_APPROVAL",
-    });
-  });
-});
