+++ packages/@cdktf/cli-core/src/test/models/deploy-machine.test.ts
@@ -5,6 +5,7 @@ import {
   deployMachine,
   extractVariableNameFromPrompt,
   terraformPtyService,
+  handleLineReceived,
 } from "../../lib/models/deploy-machine";
 
 describe("extractVariableNameFromPrompt", () => {
@@ -242,6 +243,45 @@ describe("pty events", () => {
           cwd: "",
         },
       },
+    });
+  });
+});
+
+describe("handleLineReceived", () => {
+  it("should forward the line", () => {
+    const send = jest.fn();
+    handleLineReceived(send)("foo");
+    expect(send).toHaveBeenCalledWith({
+      type: "LINE_RECEIVED",
+      line: "foo",
+    });
+  });
+
+  it("should hide the line if a variable is missing", () => {
+    const send = jest.fn();
+    const input = `
+        var.my_var
+        Enter a value:
+    `;
+    handleLineReceived(send)(input);
+    expect(send).not.toHaveBeenCalledWith({
+      type: "LINE_RECEIVED",
+      line: input,
+    });
+    expect(send).toHaveBeenCalledWith({
+      type: "VARIABLE_MISSING",
+      variableName: "my_var",
+    });
+  });
+
+  it("should detect destroy all resources in workspace message on windows", () => {
+    const send = jest.fn();
+    const input = `\u001b[25
+         l\u001b[m\n\u001b[1m\u001b[97mDo you really want to destroy all resources in workspace \\"windows-node-test\\"?   \u001b[2
+    `;
+    handleLineReceived(send)(input);
+    expect(send).toHaveBeenCalledWith({
+      type: "REQUEST_APPROVAL",
     });
   });
 });
